// src/app/bookings/enums/cancellation-reason.ts
export enum CancellationReason {
  CUSTOMER_REQUEST = 'customer_request',
  NO_SHOW = 'no_show',
  DUPLICATE = 'duplicate',
  TOUR_CANCELLED = 'tour_cancelled',
  CHANGE_OF_PLANS = 'change_of_plans',
  WEATHER_ISSUES = 'weather_issues',
  TRAVEL_DELAYS = 'travel_delays',
  HEALTH_ISSUES = 'health_issues',
  FOUND_BETTER_OFFER = 'found_better_offer',
  FINANCIAL_REASONS = 'financial_reasons',
  ACCIDENTAL_BOOKING = 'accidental_booking',
  OTHER = 'other',
}